+++ packages/credential-sdk/src/types/did/onchain/verification-method-signature.js
@@ -1,8 +1,12 @@
 import { valueBytes, ensureInstanceOf } from '../../../utils';
-import { TypedStruct } from '../../generic';
+import { TypedStruct, withProp } from '../../generic';
 import { SignatureEd25519Value } from '../../signatures';
 import { DidKeypair } from '../../../keypairs';
-import { CheqdVerificationMethodRef } from '../document/verification-method-ref';
+import {
+  CheqdMainnetVerificationMethodRef,
+  CheqdTestnetVerificationMethodRef,
+  CheqdVerificationMethodRef,
+} from '../document/verification-method-ref';
 
 class BytesSignatureEd25519Value extends SignatureEd25519Value {
   toCheqdPayload() {
@@ -25,3 +29,15 @@ export class VerificationMethodSignature extends TypedStruct {
     );
   }
 }
+
+export class CheqdTestnetVerificationMethodSignature extends withProp(
+  VerificationMethodSignature,
+  'verificationMethodId',
+  CheqdTestnetVerificationMethodRef,
+) {}
+
+export class CheqdMainnetVerificationMethodSignature extends withProp(
+  VerificationMethodSignature,
+  'verificationMethodId',
+  CheqdMainnetVerificationMethodRef,
+) {}